requirements.txt
aiohttp==3.6.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.3.0
pyTelegramBotAPI==3.7.2
python-dotenv==0.14.0
requests==2.24.0
rsa==4.6
six==1.15.0
Telethon==1.15.0
toml==0.10.1
urllib3==1.25.10
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
